pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sinc"
version = "0.1.0"
description = "Sistema Inteligente de Notificação e Captura — Supermercado Kan"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<0.116",
    "uvicorn[standard]>=0.34,<0.35",
    "sqlalchemy[asyncio]>=2.0,<3.0",
    "alembic>=1.14,<2.0",
    "asyncpg>=0.30,<0.31",
    "psycopg2-binary>=2.9,<3.0",
    "pydantic>=2.10,<3.0",
    "pydantic-settings>=2.7,<3.0",
    "redis>=5.2,<6.0",
    "structlog>=25.1,<26.0",
    "python-jose[cryptography]>=3.3,<4.0",
    "passlib[argon2]>=1.7,<2.0",
    "argon2-cffi>=23.1,<24.0",
    "python-multipart>=0.0.20,<0.1",
    "typer>=0.15,<0.16",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
    "pytest-asyncio>=0.25,<0.26",
    "httpx>=0.28,<0.29",
    "mypy>=1.14,<2.0",
    "ruff>=0.9,<0.10",
    "black>=25.1,<26.0",
    "types-passlib",
    "types-python-jose",
]

[project.scripts]
sinc = "sinc.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["sinc*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
